feat(example): 기존 공통 컴포넌트 예시 페이지 개별 페이지로 변경
