+++ apps/monitor-web/src/pages/Example/index.ts
@@ -1,0 +1,1 @@
+export { default as Example } from "./Example";
