+++ apps/monitor-web/src/pages/index.ts
@@ -5,3 +5,4 @@ export { ApiDetail } from "./ApiDetail";
 export { ApiEdit } from "./ApiEdit";
 export { ErrorDetail } from "./ErrorDetail";
 export { NotFound } from "./NotFound";
+export { Example } from "./Example";
